requirements.txt
# Albion Market Bot v3.0 Dependencies

# GUI Framework
PyQt6>=6.4.0

# OCR
pytesseract>=0.3.10
Pillow>=9.0.0

# Automation
pyautogui>=0.9.53
keyboard>=0.13.5
pynput>=1.7.6

# Data Processing
pandas>=1.5.0
openpyxl>=3.0.10
numpy>=1.23.0

# Fuzzy Matching
thefuzz>=0.20.0
python-Levenshtein>=0.20.0

# Image Processing
opencv-python>=4.7.0

# HTTP Requests (for API)
requests>=2.28.0

# Testing
pytest>=7.2.0
